+++ libavcodec/dxva2_h264.c
@@ -22,6 +22,7 @@
 
 #include "h264dec.h"
 #include "h264data.h"
+#include "h264_ps.h"
 #include "mpegutils.h"
 
 // The headers above may include w32threads.h, which uses the original
