+++ libavcodec/h264_direct.c
@@ -28,6 +28,7 @@
 #include "internal.h"
 #include "avcodec.h"
 #include "h264dec.h"
+#include "h264_ps.h"
 #include "mpegutils.h"
 #include "rectangle.h"
 #include "thread.h"
