+++ libavcodec/h264_loopfilter.c
@@ -30,6 +30,7 @@
 #include "internal.h"
 #include "avcodec.h"
 #include "h264dec.h"
+#include "h264_ps.h"
 #include "mathops.h"
 #include "mpegutils.h"
 #include "rectangle.h"
