+++ libavcodec/h264_mb.c
@@ -32,6 +32,7 @@
 #include "libavutil/intreadwrite.h"
 #include "avcodec.h"
 #include "h264dec.h"
+#include "h264_ps.h"
 #include "qpeldsp.h"
 #include "thread.h"
