+++ libavcodec/h264_parse.h
@@ -25,6 +25,7 @@
 #define AVCODEC_H264_PARSE_H
 
 #include "get_bits.h"
+#include "h264_ps.h"
 
 typedef struct H264PredWeightTable {
     int use_weight;
@@ -52,11 +53,7 @@ typedef struct H264POCContext {
     int prev_frame_num;         ///< frame_num of the last pic for POC type 1/2
 } H264POCContext;
 
-struct SPS;
-struct PPS;
-struct H264ParamSets;
-
-int ff_h264_pred_weight_table(GetBitContext *gb, const struct SPS *sps,
+int ff_h264_pred_weight_table(GetBitContext *gb, const SPS *sps,
                               const int *ref_count, int slice_type_nos,
                               H264PredWeightTable *pwt);
 
@@ -76,20 +73,20 @@ int ff_h264_check_intra_pred_mode(void *logctx, int top_samples_available,
                                   int mode, int is_chroma);
 
 int ff_h264_parse_ref_count(int *plist_count, int ref_count[2],
-                            GetBitContext *gb, const struct PPS *pps,
+                            GetBitContext *gb, const PPS *pps,
                             int slice_type_nos, int picture_structure);
 
 int ff_h264_init_poc(int pic_field_poc[2], int *pic_poc,
-                     const struct SPS *sps, H264POCContext *poc,
+                     const SPS *sps, H264POCContext *poc,
                      int picture_structure, int nal_ref_idc);
 
-int ff_h264_decode_extradata(const uint8_t *data, int size, struct H264ParamSets *ps,
+int ff_h264_decode_extradata(const uint8_t *data, int size, H264ParamSets *ps,
                              int *is_avc, int *nal_length_size,
                              int err_recognition, void *logctx);
 
 /**
  * compute profile from sps
  */
-int ff_h264_get_profile(const struct SPS *sps);
+int ff_h264_get_profile(const SPS *sps);
 
 #endif /* AVCODEC_H264_PARSE_H */
