+++ libavcodec/h264_ps.c
@@ -31,12 +31,14 @@
 #include "internal.h"
 #include "mathops.h"
 #include "avcodec.h"
-#include "h264dec.h"
 #include "h264data.h"
+#include "h264_ps.h"
 #include "golomb.h"
 
 #define MAX_LOG2_MAX_FRAME_NUM    (12 + 4)
 #define MIN_LOG2_MAX_FRAME_NUM    4
+
+#define EXTENDED_SAR       255
 
 static const AVRational pixel_aspect[17] = {
     {   0,  1 },
