+++ libavcodec/vaapi_h264.c
@@ -22,6 +22,7 @@
 
 #include "vaapi_internal.h"
 #include "h264dec.h"
+#include "h264_ps.h"
 #include "mpegutils.h"
 
 /**
